backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
watchdog==3.0.0
websockets==12.0
python-multipart==0.0.6
aiofiles==23.2.1
pyyaml==6.0.1
httpx==0.25.2